+++ src/main/java/com/example/umcmatchingcenter/controller/MatchingController.java
@@ -5,7 +5,9 @@ import com.example.umcmatchingcenter.converter.matching.MatchingConverter;
 import com.example.umcmatchingcenter.domain.Branch;
 import com.example.umcmatchingcenter.domain.Project;
 import com.example.umcmatchingcenter.domain.enums.ProjectStatus;
+import com.example.umcmatchingcenter.dto.MatchingDTO.MatchingRequestDTO;
 import com.example.umcmatchingcenter.dto.MatchingDTO.MatchingResponseDTO;
+import com.example.umcmatchingcenter.service.matchingService.MatchingCommandService;
 import com.example.umcmatchingcenter.service.matchingService.MatchingQueryService;
 import com.example.umcmatchingcenter.service.memberService.MemberQueryService;
 import com.example.umcmatchingcenter.validation.annotation.ExistMember;
@@ -17,8 +19,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
 import java.security.Principal;
@@ -32,6 +36,7 @@ public class MatchingController {
 
     private final MatchingQueryService matchingQueryService;
     private final MemberQueryService memberQueryService;
+    private final MatchingCommandService matchingCommandService;
 
     // 프로젝트 목록 조회
     @GetMapping("")
@@ -67,5 +72,20 @@ public class MatchingController {
     public ApiResponse<MatchingResponseDTO.MatchingProjectDTO> getProject(@PathVariable(name = "projectId") Long projectId){
         Project project = matchingQueryService.getProjectDetail(projectId);
         return ApiResponse.onSuccess(MatchingConverter.toMatchingProjectDetailDTO(project));
+    }
+
+    @PostMapping(value = "/projects", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
+    @Operation(summary = "아이디어 페이지 작성")
+    @ApiResponses({ // API 응답
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200",description = "OK, 성공"),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH003", description = "access 토큰을 주세요!",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH004", description = "access 토큰 만료",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH006", description = "access 토큰 모양이 이상함",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
+    })
+    @PreAuthorize("isAuthenticated()")
+    public ApiResponse<String> addMatchingProjects(@RequestPart MatchingRequestDTO.AddMatchingProjectRequestDto request, @RequestPart(required = false) MultipartFile image,
+                                                   Principal principal) {
+        matchingCommandService.addMatchingProjects(request, principal.getName(),image);
+        return ApiResponse.onSuccess("아이디어 등록에 성공하였습니다.");
     }
 }
