+++ src/main/java/com/example/umcmatchingcenter/controller/ProjectController.java
@@ -7,7 +7,6 @@ import com.example.umcmatchingcenter.domain.enums.MemberPart;
 import com.example.umcmatchingcenter.domain.enums.ProjectStatus;
 import com.example.umcmatchingcenter.dto.ProjectDTO.ProjectRequestDTO;
 import com.example.umcmatchingcenter.dto.ProjectDTO.ProjectResponseDTO;
-import com.example.umcmatchingcenter.service.projectService.ProjectCommandService;
 import com.example.umcmatchingcenter.service.projectService.ProjectQueryService;
 import com.example.umcmatchingcenter.validation.annotation.CompleteProject;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,14 +17,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.security.Principal;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -34,7 +28,6 @@ import java.util.Map;
 public class ProjectController {
 
     private final ProjectQueryService projectQueryService;
-    private final ProjectCommandService projectCommandService;
 
     // 프로젝트 목록 조회
     @GetMapping("")
@@ -68,21 +61,6 @@ public class ProjectController {
     public ApiResponse<ProjectResponseDTO.ProjectDTO> getProject(@CompleteProject @PathVariable(name = "projectId") Long projectId) {
         Project project = projectQueryService.getProjectDetail(projectId);
         return ApiResponse.onSuccess(ProjectConverter.toProjectDetailDTO(project));
-    }
-
-    @PostMapping(value = "/projects", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
-    @Operation(summary = "아이디어 페이지 작성")
-    @ApiResponses({ // API 응답
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200",description = "OK, 성공"),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH003", description = "access 토큰을 주세요!",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH004", description = "access 토큰 만료",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH006", description = "access 토큰 모양이 이상함",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
-    })
-    @PreAuthorize("isAuthenticated()")
-    public ApiResponse<String> addProjects(@RequestPart ProjectRequestDTO.AddProjectRequestDto request, @RequestPart(required = false) MultipartFile image,
-                                                                  Principal principal) {
-        projectCommandService.addProject(request, principal.getName(),image);
-        return ApiResponse.onSuccess("아이디어 등록에 성공하였습니다.");
     }
 }
